requirements.txt
python-telegram-bot==10.1.0
requests==2.19.1
requests-cache==0.4.13